streaming-pipeline/pyproject.toml
[tool.poetry]
name = "streaming-pipeline"
version = "0.1.0"
description = ""
authors = ["Massyl GH <massyl-6@hotmail.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
bytewax = "^0.21.1"
python-dotenv = "^1.1.0"
pydantic = "^2.11.2"
qdrant-client = "^1.13.3"
langchain = "^0.3.23"
transformers = "^4.51.1"
torch = "^2.6.0"
fire = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
